requirements.txt
# Name                    Version                   Build  Channel
biopython                 1.73             py36h1de35cc_0  
dgl                       0.4.1                    pypi_0    pypi
networkx                  2.4                        py_0  
numpy                     1.16.3           py36hacdab7b_0  
openbabel                 2.4.1                    py36_5    openbabel
pandas                    0.24.2           py36h0a44026_0  
pytorch                   1.1.0                   py3.6_0    pytorch
scipy                     1.2.1            py36h1410ff5_0  
seaborn                   0.9.0                    py36_0  
